+++ src/core/model/shapes/trim.cpp
@@ -24,24 +24,54 @@ bool glaxnimate::model::Trim::process_collected() const
 
 static void chunk_start(const glaxnimate::math::bezier::Bezier& in, glaxnimate::math::bezier::Bezier& out, const glaxnimate::math::bezier::LengthData::SplitInfo& split, int max = -1)
 {
+    using namespace glaxnimate::math::bezier;
+
     if ( max == -1 )
         max = in.closed_size();
 
+    // empty
     if ( split.ratio == 0 && split.index == 0 && max == in.closed_size() )
     {
         out = in;
         return;
     }
 
-    if ( split.ratio < 1 )
-        out.push_back(in.split_segment_point(split.index, split.ratio));
+    int index = split.index;
 
-    for ( int i = split.index + 1; i < max; i++ )
+    // gotta split mid segment
+    if ( split.ratio < 1 && split.ratio > 0 )
+    {
+
+        auto split_points = CubicBezierSolver<QPointF>(in.segment(split.index)).split(split.ratio);
+        out.push_back(Point(
+            split_points.first[3],
+            split_points.first[2],
+            split_points.second[1],
+            Smooth
+        ));
+        index += 1;
+
+        //we use the tangents from the split for the next point
+        if ( index < max )
+        {
+            out.push_back(Point(
+                split_points.second[3],
+                split_points.second[2],
+                in[index].tan_out,
+                in[index].type
+            ));
+            index += 1;
+        }
+    }
+
+
+    for ( int i = index; i < max; i++ )
         out.push_back(in[i]);
 }
 
 static void chunk_end(const glaxnimate::math::bezier::Bezier& in, glaxnimate::math::bezier::Bezier& out, const glaxnimate::math::bezier::LengthData::SplitInfo& split, int min = 0)
 {
+    using namespace glaxnimate::math::bezier;
 
     if ( split.ratio == 1 && min == 0 )
     {
@@ -52,8 +82,22 @@ static void chunk_end(const glaxnimate::math::bezier::Bezier& in, glaxnimate::ma
     for ( int i = min; i <= split.index; i++ )
         out.push_back(in[i]);
 
+    // gotta split mid segment
     if ( split.ratio > 0 )
-        out.push_back(in.split_segment_point(split.index, split.ratio));
+    {
+        auto split_points = CubicBezierSolver<QPointF>(in.segment(split.index)).split(split.ratio);
+
+        // adjust tangents for the pevious point
+        if ( !out.empty() )
+            out[out.size()-1].tan_out = split_points.first[1];
+
+        out.push_back(Point(
+            split_points.first[3],
+            split_points.first[2],
+            split_points.second[1],
+            Smooth
+        ));
+    }
 }
 
 glaxnimate::math::bezier::MultiBezier glaxnimate::model::Trim::process(glaxnimate::model::FrameTime t, const math::bezier::MultiBezier& mbez) const
@@ -62,13 +106,19 @@ glaxnimate::math::bezier::MultiBezier glaxnimate::model::Trim::process(glaxnimat
         return {};
 
     auto offset = this->offset.get_at(t);
-    auto start = math::fmod(this->start.get_at(t) + offset, 1.f);
-    auto end = math::fmod(this->end.get_at(t) + offset, 1.f);
-    if ( end == 0 )
+    auto start_value = this->start.get_at(t);
+    auto start = math::fmod(start_value + offset, 1.f);
+    auto end_value = this->end.get_at(t);
+    auto end = math::fmod(end_value + offset, 1.f);
+    if ( qFuzzyIsNull(end) && end_value > start_value )
         end = 1;
 
     if ( start == 0 && end == 1 )
         return mbez;
+
+    // Same value
+    if ( math::abs(start * 1000 - end * 1000) < 1 )
+        return {};
 
     const int length_steps = 5;
 
@@ -92,9 +142,45 @@ glaxnimate::math::bezier::MultiBezier glaxnimate::model::Trim::process(glaxnimat
             auto single_start_data = start_data.child_split();
             auto single_end_data = end_data.child_split();
             math::bezier::Bezier b;
-            chunk_start(mbez.beziers()[start_data.index], b, single_start_data, single_end_data.index);
-            int end_min = qMax(single_end_data.index, single_start_data.index + 1);
-            chunk_end(mbez.beziers()[start_data.index], b, single_end_data, end_min);
+
+            /**
+             * Same bezier segment
+             * [ seg[0] ... seg[single_start=single_end]  ... seg[m] ]
+             *              aaaaa|BBBBBBBBBBBBBBBB|ccccc
+             */
+            if ( single_start_data.index == single_end_data.index )
+            {
+                const auto& in = mbez.beziers()[start_data.index];
+                int index = single_start_data.index;
+                // split the segment at start
+                qreal ratio_start = single_start_data.ratio;
+                auto truncated_segment = math::bezier::CubicBezierSolver<QPointF>(in.segment(index)).split(ratio_start).second;
+                // find the end ratio for the truncated segment and split it there
+                qreal ratio_end = (single_end_data.ratio - ratio_start) / (1-ratio_start);
+                auto result = math::bezier::CubicBezierSolver<QPointF>(truncated_segment).split(ratio_end).first;
+                // add to the bezier
+                b.push_back(math::bezier::Point(
+                    result[0],
+                    result[0],
+                    result[1],
+                    math::bezier::Corner
+                ));
+                b.push_back(math::bezier::Point(
+                    result[3],
+                    result[2],
+                    result[3],
+                    math::bezier::Corner
+                ));
+            }
+            else
+            {
+                int start_max = single_end_data.index;
+                if ( single_end_data.index == single_start_data.index + 1 )
+                    start_max += 1;
+                chunk_start(mbez.beziers()[start_data.index], b, single_start_data, start_max);
+                int end_min = qMax(start_max, single_start_data.index + 1);
+                chunk_end(mbez.beziers()[start_data.index], b, single_end_data, end_min);
+            }
             out.beziers().push_back(b);
             return out;
         }