+++ src/gui/main.cpp
@@ -8,6 +8,7 @@
 #include "cli.hpp"
 #include "app_info.hpp"
 #include "io/io_registry.hpp"
+#include "io/lottie/lottie_html_format.hpp"
 
 #include "widgets/dialogs/glaxnimate_window.hpp"
 
@@ -49,7 +50,10 @@ int main(int argc, char *argv[])
 
     app.initialize();
 
-    gui::GlaxnimateWindow window(!args.has_flag("default-ui"), args.has_flag("debug"));
+    bool debug = args.has_flag("debug");
+    if ( debug )
+        io::IoRegistry::instance().register_object(std::make_unique<io::lottie::LottieHtmlFormat>());
+    gui::GlaxnimateWindow window(!args.has_flag("default-ui"), debug);
     sc.finish(&window);
     window.show();